Add files via upload
Dataset recording people to invest in each other in a way that is financially and socially rewarding. On loans, borrowers list loan requests between $ 2,000 and $ 35,000, and individual investors invest as little as $ 25 in each loan listing they select. Prosper handles the servicing of the loan on behalf of the matched borrowers and investors

Skills and Tools

Descriptive statistics, preliminary wrangling, univariate analysis, bivariate analysis, multivariate analysis, feature engineering